refactor: devide Main.java into two files
- Main.java 파일의 크기가 커서 가독성이 떨어졌기 때문에, main method와 constant가
  있는 Main.java 파일과, Deque 로직이 있는 SooDeque.java 파일로 구분했습니다.
